dtype follows input data

+++ test_plsa.py
@@ -67,16 +67,16 @@ class TestUtility(unittest.TestCase):
 class TestPLSA(unittest.TestCase):
     @classmethod
     def setUpClass(self):
-        data1 = torch.tensor([[3, 0, 0], [1, 0, 0], [0, 1, 1]])
+        data1 = torch.tensor([[3.0, 0, 0], [1, 0, 0], [0, 1, 1]])
         self.data1 = data1
         nclass = 2
         self.plsa1 = plsa.PLSA(data1, nclass)
         data2 = torch.tensor(
             [
-                [[3, 0, 0], [1, 0, 0], [0, 0, 0]],
+                [[3.0, 0, 0], [1, 0, 0], [0, 0, 0]],
                 [[3, 0, 0], [1, 0, 0], [0, 1, 1]],
                 [[0, 0, 0], [0, 0, 0], [0, 1, 1]]
-            ])
+            ], dtype=torch.float64)
         nclass = 2
         self.plsa2 = plsa.PLSA(data2, nclass)
 
@@ -98,8 +98,7 @@ class TestPLSA(unittest.TestCase):
 
     def test_em_algorithm_dim0(self):
         # Example of document likelihood estimation with given p(z) and p(x1|z)
-        self.plsa1.data = torch.tensor([[4, 0, 0], [1, 5, 0], [1, 2, 1]],
-                                       dtype=torch.float)
+        self.plsa1.data = torch.tensor([[4.0, 0, 0], [1, 5, 0], [1, 2, 1]])
         self.plsa1.reset(inds_fixed=[-1, 0])
         self.plsa1.em_algorithm(10, inds_fixed=[-1, 0])
         ans = [torch.tensor([[0.75, 0.25, 0.00], [0.00, 0.00, 1.00]]),
